user save edit added

+++ MANDRIN/mandrin-core/src/main/java/com/mandrin/core/domain/AbstractTrackableEntity.java
@@ -18,7 +18,7 @@ public class AbstractTrackableEntity implements Serializable{
 	@Version
 	private Integer version;
 
-	public long getVersion() {
+	public Integer getVersion() {
 		return version;
 	}
